에디터에서 블록 가로 스크롤 (#1235)
가로로 넘치는 블록이 자연스럽게 가로 스크롤이 되도록 함

- 블록 width는 이제 720px이 아니라 full임
- 대신 패딩을 넣어서 content box가 720px이 되도록 함
- 블록에 overflowX: auto 적용
- block-selection decoration이 여전히 똑같이 보이도록 수정함
- floating menu가 여전히 똑같은 위치에 보이도록 수정함 + 블록을 가로로 스크롤해도 적절한 위치에 보이도록 블록의 scrollLeft를 이용함
- 사이드바 위로 floating menu가 올라오지 않도록 사이드바에 zIndex 50 줌
- 바깥으로 튀어나온 table의 행/열 핸들과 행/열 추가 버튼이 여전히 잘리지 않고 잘 보일 수 있도록 함
- table node를 선택하면 행/열 핸들과 행/열 추가 버튼이 보이지 않도록 함 (잘리지 않게 보이게 하려면 까다로움, 노션과 동일한 동작임)
- 행/열 핸들과 행/열 추가 버튼 컴포넌트를 분리함
- 마지막 행 핸들에 '아래로 이동' 메뉴가 보이는 문제 디버그
